- fix compilation error.
